backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
elevenlabs==0.2.24
google-cloud-texttospeech==2.14.1
uvicorn==0.34.0
azure-cognitiveservices-vision-computervision==0.9.0
Pillow==10.0.0
